Adds `(archs4)_series_status` functions to check for missing samples.
This function is made available to the user so that she can query GEO series
and check if the series is missing any samples.

This is also (importantly) used by the `(archs4_)sample_info` and `as.DGEList`
functions. When series id's are used in these queries, they are expanded such
that the returned objects include data from the sampels in the series. If a
series is missing samples, these functions will raise a warning to notify the
user.